Added arch filter to prevent app crash
On Android when using old Battle Breakers armeabi-v7a on arm64 device, if the arm64 library is inside the APK, the system will proceed to load it rather than the armeabi-v7a with the translation layer. This seems to be an issue since the app just crashes (probably due to the hook placed by the arm64 lib inside the "arm64 translated" libUE4.so). This commit fixes it by seeking for libUE4.so 's architecture.